// superset-frontend/src/pages/DashboardTemplates/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

export interface DashboardTemplate {
  id: number | null; // null for "Start from blank"
  uuid: string;
  dashboard_title: string;
  slug?: string;
  template_description?: string;
  template_category?: string | null;
  is_featured_template: boolean;
  is_template: boolean;
  template_thumbnail_url?: string | null;
  template_tags: string[];
  template_context?: string;
}

export interface DashboardTemplateGalleryProps {
  templates: DashboardTemplate[];
  loading: boolean;
  onSelectTemplate: (template: DashboardTemplate | null) => void;
}